+++ cost/aws/rightsize_rds_instances/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## v5.8.3
+
+- Updated meta policy code to use newer Flexera API. Functionality unchanged.
+
 ## v5.8.2
 
 - Added `doc_link` field to policy template metadata for future UI enhancements. Functionality unchanged.