+++ cost/azure/long_stopped_instances/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## v6.0.3
+
+- Updated meta policy code to use newer Flexera API. Functionality unchanged.
+
 ## v6.0.2
 
 - Added `doc_link` field to policy template metadata for future UI enhancements. Functionality unchanged.