+++ cost/azure/rightsize_sql_instances/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## v6.0.4
+
+- Updated meta policy code to use newer Flexera API. Functionality unchanged.
+
 ## v6.0.3
 
 - Fixed bug that was causing Maximum, Minimum to not get selected correctly