+++ operational/azure/azure_long_running_instances/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## v6.1.5
+
+- Updated meta policy code to use newer Flexera API. Functionality unchanged.
+
 ## v6.1.4
 
 - Updated API requests to use newer Flexera API. Functionality unchanged.